Editorial: Define + use StringToNumber
Formerly, the procedure for applying ToNumber to the String type
was spread over three widely-separated prose paragraphs
in two different clauses.

This commit brings all that together, expresses it as an actual algorithm,
and gives it the name StringToNumber.

Also, the definitions of the syntax-direction operation
NumericValue mostly just delegated to a prose description.
This commit replaces that with actual algorithms,
similar to that for StringToNumber.