Editorial: Replace "integer index" with "Uint53-string"
Also, define "Uint53-indexed" and use it to replace "integer-indexed".

(Remaining occurrences of "integer index" and "integer-indexed"
are using the phrases not as the terms were defined.
That is, they are talking about an integer value
that's an index into some sequence.)

(Oh, and I'm also leaving the term "Integer-Indexed exotic object".)