Tweak grammatical parameters in 'FieldDefinition' production
First, in a RHS, just 'In' is invalid, you need either +In or ~In.
You probably want +In.

Second, [~Yield, ~Await] is valid, but I suspect it's wrong.
Every other use of Initializer is happy with [?Yield, ?Await],
and I don't see a reason why this particular one would need to force [~Yield, ~Await].

Please check.